+++ lib/live_data_source.dart
@@ -1,6 +1,7 @@
 import 'dart:async';
 import 'live_data.dart';
 
+@deprecated
 class LiveDataSource<T> extends LiveData<T> {
   DataSourceInterface? _dataSourceInterface;
 
@@ -41,6 +42,7 @@ class LiveDataSource<T> extends LiveData<T> {
   }
 }
 
+@deprecated
 abstract class DataSourceInterface<T> {
   Future<void> onValueUpdated(T value, bool hasChange);
 
@@ -57,6 +59,7 @@ DataSourceInterface<T> createDataSourceInterface<T>({
   );
 }
 
+@deprecated
 class _DataSourceInterface<T> extends DataSourceInterface<T> {
   Future<T> Function()? loadValueAction;
   Future<void> Function(T value, bool hasChange)? onValueUpdatedAction;